Extract node resources packages
